fix: Correct type conversion in `PerformanceDaoTest`
The `toEpochDays()` method returns a `Long`, which was causing a type mismatch. This commit fixes the issue by explicitly converting the result to an `Int`.

+++ app/src/androidTest/kotlin/com/looker/kenko/PerformanceDaoTest.kt
@@ -228,7 +228,7 @@ class PerformanceDaoTest {
     private fun performanceOf(sets: Map<LocalDate, Set>): Performance = sets
         .map { (date, set) ->
             RatingWrapper(
-                date = date.toEpochDays(),
+                date = date.toEpochDays().toInt(),
                 rating = set.rating.value,
             )
         }
@@ -240,7 +240,7 @@ class PerformanceDaoTest {
             .groupBy({ it.first }) { it.second }
             .map { (date, sets) ->
                 RatingWrapper(
-                    date = date.toEpochDays(),
+                    date = date.toEpochDays().toInt(),
                     rating = sets.sumOf { it.rating.value.toDouble() }.toFloat(),
                 )
             }
